+++ app/src/main/AndroidManifest.xml
@@ -13,16 +13,16 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity android:name=".MainActivity" />
-        <activity android:name=".PreviewActivity" />
-        <activity android:name=".AboutActivity" />
-        <activity android:name=".news.NewsListActivity">
+        <activity android:name=".Activity.PreviewActivity" />
+        <activity android:name=".Activity.AboutActivity" />
+        <activity android:name=".news.NewsListActivity"/>
+        <activity android:name=".Activity.IntroActivity" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".news.NewsDetailActivity" />
     </application>
 
 </manifest>
