+++ app/src/main/java/ru/evgeniy/androidacademy/Activity/IntroActivity.java
@@ -1,0 +1,47 @@
+package ru.evgeniy.androidacademy.Activity;
+
+import android.content.Intent;
+import android.content.SharedPreferences;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+import java.util.concurrent.TimeUnit;
+
+import io.reactivex.Completable;
+import io.reactivex.disposables.CompositeDisposable;
+import io.reactivex.disposables.Disposable;
+import ru.evgeniy.androidacademy.R;
+import ru.evgeniy.androidacademy.data.SharedPref;
+import ru.evgeniy.androidacademy.news.NewsListActivity;
+
+public class IntroActivity extends AppCompatActivity {
+
+    SharedPref Storage;
+    private CompositeDisposable compositeDisposable = new CompositeDisposable();
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        Storage = new SharedPref(this);
+        if (Storage.needToShowIntro()) {
+            setContentView(R.layout.activity_intro);
+            Disposable disposable = Completable.complete()
+                    .delay(10, TimeUnit.SECONDS)
+                    .subscribe(this::startSecondActivity);
+            compositeDisposable.add(disposable);
+        } else {
+            startSecondActivity();
+        }
+    }
+
+    private void startSecondActivity() {
+        startActivity(new Intent(this, NewsListActivity.class));
+        finish();
+    }
+
+    @Override
+    protected void onStop() {
+        super.onStop();
+        compositeDisposable.dispose();
+    }
+}
