+++ app/src/main/java/ru/evgeniy/androidacademy/MainActivity.java
@@ -5,10 +5,10 @@ import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.text.Editable;
 import android.text.TextWatcher;
-import android.util.Patterns;
-import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+
+import ru.evgeniy.androidacademy.Activity.PreviewActivity;
 
 public class MainActivity extends AppCompatActivity {
